Refactor storage client
